add option to enter code manually on login
This fixes a problem with corporate email security systems that follow links within email, therefore triggering the token so it can't be used to actually log in.

Users now have the option to get an email containing the code outside the URL, then manually enter it in a form.

Fixes #112